Extract rewriteAndWrite() test helper to fix SonarCloud duplication gate
fc-generate-aux-config.test.mjs duplicated the import-rewriting helper
already present in connection-send-queue.test.mjs almost verbatim,
tripping the new-code duplication gate. Move it to a shared factory in
tests/helpers/rewriteAndWrite.mjs and have this file use that instead of
its own copy.
